+++ test/shortest_path_test.dart
@@ -5,11 +5,15 @@
 import 'package:graphs/graphs.dart';
 import 'package:test/test.dart';
 
+import 'utils/utils.dart';
+
 Matcher _throwsAssertionError(messageMatcher) =>
     throwsA(const TypeMatcher<AssertionError>()
         .having((ae) => ae.message, 'message', messageMatcher));
 
-String _identity(int input) => input.toString();
+int _xKey(X input) => input.value;
+
+T _identity<T>(T input) => input;
 
 void main() {
   const graph = <int, List<int>>{
@@ -21,14 +25,15 @@ void main() {
     6: [7],
   };
 
+  List<int> getValues(int key) => graph[key] ?? [];
+
+  List<X> getXValues(X key) =>
+      graph[key.value]?.map((v) => X(v))?.toList() ?? [];
+
   test('null `start` throws AssertionError', () {
-    expect(
-        () => shortestPath<String, int>(
-            null, 1, _identity, (input) => graph[input] ?? []),
+    expect(() => shortestPath(null, 1, _identity, getValues),
         _throwsAssertionError('`start` cannot be null'));
-    expect(
-        () => shortestPaths<String, int>(
-            null, _identity, (input) => graph[input] ?? []),
+    expect(() => shortestPaths(null, _identity, getValues),
         _throwsAssertionError('`start` cannot be null'));
   });
 
@@ -58,24 +63,32 @@ void main() {
   });
 
   void _singlePathTest(int from, int to, List<int> expected) {
-    test('$from -> $to should be $expected', () {
+    test('$from -> $to should be $expected (mapped)', () {
       expect(
-          shortestPath<String, int>(
-              from, to, _identity, (input) => graph[input] ?? []),
+          shortestPath<int, X>(X(from), X(to), _xKey, getXValues)
+              ?.map((x) => x.value),
           expected);
+    });
+
+    test('$from -> $to should be $expected', () {
+      expect(shortestPath<int, int>(from, to, _identity, getValues), expected);
     });
   }
 
-  void _pathsTest(
-      int from, Map<String, List<int>> expected, List<int> nullPaths) {
+  void _pathsTest(int from, Map<int, List<int>> expected, List<int> nullPaths) {
+    test('paths from $from (mapped)', () {
+      final result = shortestPaths<int, X>(X(from), _xKey, getXValues)
+          .map((k, v) => MapEntry(k, v.map((x) => x.value).toList()));
+      expect(result, expected);
+    });
+
     test('paths from $from', () {
-      final result = shortestPaths<String, int>(
-          from, _identity, (input) => graph[input] ?? []);
+      final result = shortestPaths<int, int>(from, _identity, getValues);
       expect(result, expected);
     });
 
     for (var entry in expected.entries) {
-      _singlePathTest(from, int.parse(entry.key), entry.value);
+      _singlePathTest(from, entry.key, entry.value);
     }
 
     for (var entry in nullPaths) {
@@ -84,26 +97,26 @@ void main() {
   }
 
   _pathsTest(1, {
-    '5': [5],
-    '3': [2, 3],
-    '8': [5, 8],
-    '1': [],
-    '2': [2],
-    '4': [2, 3, 4],
+    5: [5],
+    3: [2, 3],
+    8: [5, 8],
+    1: [],
+    2: [2],
+    4: [2, 3, 4],
   }, [
     6,
     7,
   ]);
 
   _pathsTest(6, {
-    '7': [7],
-    '6': [],
+    7: [7],
+    6: [],
   }, [
     1,
   ]);
-  _pathsTest(7, {'7': []}, [1, 6]);
+  _pathsTest(7, {7: []}, [1, 6]);
 
-  _pathsTest(42, {'42': []}, [1, 6]);
+  _pathsTest(42, {42: []}, [1, 6]);
 
   _singlePathTest(1, null, null);
 }